Moved all osg methods and attributes from Node and Edge to OsgNode and OsgEdge
