+++ include/Data/OsgNode.h
@@ -6,6 +6,7 @@
 
 namespace Data {
 
+typedef osg::TemplateIndexArray<unsigned int, osg::Array::UIntArrayType,4,1> ColorIndexArray;
 
 class OsgNode : public osg::Geode
 {
@@ -18,7 +19,6 @@ public:
 	    *  \return osg::ref_ptr node label
 	    */
 	static osg::ref_ptr<osg::Drawable> createLabel( const float& scale, QString name );
-
 
 	/**
 	    *  \fn private static  createNode(const float & scale, osg::StateSet* bbState)
@@ -46,33 +46,195 @@ public:
 	    */
 	static osg::ref_ptr<osg::StateSet> createStateSet( Data::Type* type = 0 );
 
-	/**
-	    *  osg::ref_ptr focusedSquare
-	    *  \brief Focused square drawable
-	    */
-	osg::ref_ptr<osg::Drawable> focusedSquare;
+    /**
+         * \fn public constant targetPosition
+         * \brief Gets target position of a node.
+         * \return target position
+         * Returned target position IS NOT multiplied by the graph scale.
+         */
+    osg::Vec3f targetPosition() const;
 
+    osg::Vec3f getTargetPosition() const;
 
-	/**
-	    *  osg::ref_ptr label
-	    *  \brief Label drawable
-	    */
-	osg::ref_ptr<osg::Drawable> label;
+    /**
+         * \fn public constant targetPositionConstRef
+         * \brief Gets target position of a node.
+         * \return target position
+         * Returned target position IS NOT multiplied by the graph scale.
+         */
+    const osg::Vec3f& targetPositionConstRef() const;
 
+    /**
+         * \fn public setTargetPosition(const osg::Vec3f &position)
+         * \brief Sets target position of a node.
+         * \param position target position
+         * Target position being set MUST NOT BE multiplied by the graph scale.
+         */
+    void setTargetPosition( const osg::Vec3f& position );
 
-	/**
-	    *  osg::ref_ptr square
-	    *  \brief Square drawable
-	    */
-	osg::ref_ptr<osg::Drawable> square;
+    /**
+         * \fn public constant restrictedTargetPosition
+         * \brief Gets restricted target position of a node.
+         * \return restricted target position
+         * Returned restricted target position IS NOT multiplied by the graph scale.
+         */
+    osg::Vec3f restrictedTargetPosition() const;
 
+    /**
+         * \fn public constant restrictedTargetPositionConstRef
+         * \brief Gets restricted target position of a node.
+         * \return restricted target position
+         * Returned restricted target position IS NOT multiplied by the graph scale.
+         */
+    const osg::Vec3f& restrictedTargetPositionConstRef() const;
 
+    /**
+         * \fn public setRestrictedTargetPosition(const osg::Vec3f &position)
+         * \brief Sets restricted target position of a node.
+         * \param position restricted target position
+         * Restricted target position being set MUST NOT BE multiplied by the graph scale.
+         */
+    void setRestrictedTargetPosition( const osg::Vec3f& position );
 
-	/**
-	    *  osg::Vec4 color
-	    *  \brief Color of the Node
-	    */
-	osg::Vec4 color;
+    /**
+        *  \fn public  getCurrentPosition(bool calculateNew = false, float interpolationSpeed = 1.0f)
+        *  \brief Returns node actual position
+        *  \param      calculateNew    If true, new position will be calculated through interpolation
+        *  \param      float   interpolation speed
+        *  \return osg::Vec3f actual position
+        *  returned currentPosition IS already multiplied by the graph scale
+        */
+    osg::Vec3f getCurrentPosition( bool calculateNew = false, float interpolationSpeed = 1.0f );
+
+    /**
+        *  \fn inline public  setCurrentPosition(osg::Vec3f val)
+        *  \brief Sets current node position
+        *  \param   val  current node position
+        *  currentPosition being set MUST BE multiplied by the graph scale
+        */
+    void setCurrentPosition( osg::Vec3f val )
+    {
+        currentPosition.set( val );
+    }
+
+    /**
+        *  \fn inline public  setForce(osg::Vec3f v)
+        *  \brief Sets force of node
+        *  \param      v  Force vector
+        */
+    void setForce( osg::Vec3f v )
+    {
+        force = v;
+    }
+
+    /**
+        *  \fn inline public constant  getForce
+        *  \brief Gets force of node
+        *  \return osg::Vec3f Force vector
+        */
+    osg::Vec3f getForce() const
+    {
+        return force;
+    }
+
+    /**
+        *  \fn inline public  addForce(bool fixed)
+        *  \brief Adds force V to node force
+        *  \param       v  Force V
+        */
+    void addForce( osg::Vec3f v )
+    {
+        force += v;
+    }
+
+    /**
+        *  \fn inline public  resetForce
+        *  \brief Sets nodes force to zero value.
+        */
+    void resetForce()
+    {
+        force = osg::Vec3f( 0,0,0 );
+    }
+
+    /**
+        *  \fn inline public  setVelocity(osg::Vec3f v)
+        *  \brief Sets node force for next iteration
+        *  \param    v  Force in actual iteration
+        */
+    void setVelocity( osg::Vec3f v )
+    {
+        velocity = v;
+    }
+
+    /**
+        *  \fn inline public  resetVelocity
+        *  \brief Reset node force for next iteration
+        */
+    void resetVelocity()
+    {
+        velocity = osg::Vec3( 0,0,0 );
+    }
+
+    /**
+        *  \fn inline public  setColor(osg::Vec4 color)
+        *  \brief Sets default node color
+        *  \param     color   default color
+        */
+    void setColor( osg::Vec4 color )
+    {
+        this->color = color;
+
+        if ( !selected ) {
+            setDrawableColor( 0, color );
+        }
+    }
+
+    void setInvisible()
+    {
+        setColor( osg::Vec4( 0,0,0,0 ) );
+    }
+
+    void setParentBall( osg::Sphere* val )
+    {
+        parentBall = val;
+    }
+
+    osg::Sphere* getParentBall()
+    {
+        return parentBall;
+    }
+
+    void setBall( osg::Geode* val )
+    {
+        ball = val;
+    }
+
+    /**
+        *  \fn inline public  setSelected(bool selected) and  color of selected node
+        *  \brief Sets node picked state
+        *  \param     selected     picked state
+        */
+    void setSelected( bool selected )
+    {
+        if ( selected ) {
+            setDrawableColor( 0, osg::Vec4( 0.0f, 0.1f, 1.0f, 1.0f ) );    // color of selected node
+        }
+        else {
+            setDrawableColor( 0, color );
+        }
+
+        this->selected = selected;
+    }
+
+    /**
+        *  \fn inline public constant  isSelected
+        *  \brief Returns if the Node is selected
+        *  \return bool true, if the Node is selected
+        */
+    bool isSelected() const
+    {
+        return selected;
+    }
 
 	/**
 	    *  \fn inline public constant  getColor
@@ -85,13 +247,6 @@ public:
 	}
 
 	/**
-	    *  osg::Vec3f velocity
-	    *  \brief Size of node force in previous iteration
-	    */
-	osg::Vec3f velocity;
-
-
-	/**
 	    *  \fn inline public constant  getVelocity
 	    *  \brief Sets node force for next iteration.
 	    *  \return osg::Vec3f Node force
@@ -101,20 +256,118 @@ public:
 		return velocity;
 	}
 
-	/**
-	    *  osg::Sphere nested ball
-	    *  \brief
-	    */
-	osg::Sphere* parentBall;
+    osg::ref_ptr<osg::AutoTransform> getOutBall()
+    {
+        return outBall;
+    }
 
-	osg::Geode* ball;
-
-	osg::ref_ptr<osg::AutoTransform> outBall;
+    void setOutBall( osg::ref_ptr<osg::AutoTransform> val )
+    {
+        outBall = val;
+    }
 
 	osg::Geode* getBall()
 	{
 		return ball;
 	}
+
+    /**
+        *  \fn private  setDrawableColor(int pos, osg::Vec4 color)
+        *  \brief Sets drawble color
+        *  \param     pos     drawable position
+        *  \param     color     drawable color
+        */
+    void setDrawableColor( int pos, osg::Vec4 color );
+
+    /**
+        *  \fn inline public  setUsingInterpolation(bool val)
+        *  \brief Sets if the Node is using interpolation or not
+        *  \param      val
+        */
+    void setUsingInterpolation( bool val )
+    {
+        usingInterpolation = val;
+    }
+
+protected:
+    /**
+        *  osg::ref_ptr focusedSquare
+        *  \brief Focused square drawable
+        */
+    osg::ref_ptr<osg::Drawable> focusedSquare;
+
+    /**
+        *  osg::ref_ptr label
+        *  \brief Label drawable
+        */
+    osg::ref_ptr<osg::Drawable> label;
+
+    /**
+        *  osg::ref_ptr square
+        *  \brief Square drawable
+        */
+    osg::ref_ptr<osg::Drawable> square;
+
+    /**
+        *  osg::Vec3f velocity
+        *  \brief Size of node force in previous iteration
+        */
+    osg::Vec3f velocity;
+
+    /**
+        *  osg::Vec4 color
+        *  \brief Color of the Node
+        */
+    osg::Vec4 color;
+
+    /**
+         * osg::Vec3f mTargetPosition
+         * \brief Target position of a node.
+         */
+    osg::Vec3f mTargetPosition;
+
+    /**
+      *  osg::Vec3f force
+      *  \brief Node force
+      */
+    osg::Vec3f force;
+
+    /**
+         * osg::Vec3f mRestrictedTargetPosition
+         * \brief Restricted target position of a node.
+         */
+    osg::Vec3f mRestrictedTargetPosition;
+
+    /**
+        *  osg::Vec3f currentPosition
+        *  \brief node current position
+        */
+    osg::Vec3f currentPosition;
+
+    osg::Vec4 colorOfNode;
+
+    /**
+        *  bool selected
+        *  \brief node picked state
+        */
+    bool selected;
+
+    /**
+        *  bool usingInterpolation
+        *  \brief node interpolation usage
+        */
+    bool usingInterpolation;
+
+    /**
+        *  osg::Sphere nested ball
+        *  \brief
+        */
+    osg::Sphere* parentBall;
+
+    osg::Geode* ball;
+
+    osg::ref_ptr<osg::AutoTransform> outBall;
+
 };
 }
 #endif // OSGNODE_H
