+++ src/Data/OsgNode.cpp
@@ -1,4 +1,18 @@
 #include "Data/OsgNode.h"
+#include "Data/Node.h"
+#include "Data/Graph.h"
+#include "Data/Cluster.h"
+
+#include "Util/ApplicationConfig.h"
+
+#include <osg/Geometry>
+#include <osg/Depth>
+#include <osg/CullFace>
+#include <osgText/FadeText>
+
+#include <QTextStream>
+
+#include "Data/OsgNode.h"
 
 osg::ref_ptr<osg::Drawable> Data::OsgNode::createLabel( const float& scale, QString name )
 {
@@ -136,6 +150,64 @@ osg::ref_ptr<osg::StateSet> Data::OsgNode::createStateSet( Data::Type* type )
 	stateSet->setAttributeAndModes( cull, osg::StateAttribute::ON );
 
 	return stateSet;
+}
+
+osg::Vec3f Data::OsgNode::getCurrentPosition( bool calculateNew, float interpolationSpeed )
+{
+    //zisime aktualnu poziciu uzla v danom okamihu
+    if ( calculateNew ) {
+        float graphScale = Util::ApplicationConfig::get()->getValue( "Viewer.Display.NodeDistanceScale" ).toFloat();
+
+        //osg::Vec3 directionVector = osg::Vec3(targetPosition.x(), targetPosition.y(), targetPosition.z()) * graphScale - currentPosition;
+        osg::Vec3 directionVector = osg::Vec3( mRestrictedTargetPosition.x(), mRestrictedTargetPosition.y(), mRestrictedTargetPosition.z() ) * graphScale - currentPosition;
+        this->currentPosition = osg::Vec3( directionVector * ( usingInterpolation ? interpolationSpeed : 1 ) + this->currentPosition );
+    }
+
+    return osg::Vec3( this->currentPosition );
+}
+
+void Data::OsgNode::setDrawableColor( int pos, osg::Vec4 color )
+{
+    //nastavenie farby uzla
+    osg::Geometry* geometry  = dynamic_cast<osg::Geometry*>( this->getDrawable( pos ) );
+
+    if ( geometry != NULL ) {
+        osg::Vec4Array* colorArray =  dynamic_cast<osg::Vec4Array*>( geometry->getColorArray() );
+
+        colorArray->pop_back();
+        colorArray->push_back( color );
+    }
+}
+
+
+
+osg::Vec3f Data::OsgNode::getTargetPosition() const
+{
+    return mTargetPosition;
+}
+osg::Vec3f Data::OsgNode::targetPosition() const
+{
+    return mTargetPosition;
+}
+const osg::Vec3f& Data::OsgNode::targetPositionConstRef() const
+{
+    return mTargetPosition;
+}
+void Data::OsgNode::setTargetPosition( const osg::Vec3f& position )
+{
+    mTargetPosition = position;
+}
+osg::Vec3f Data::OsgNode::restrictedTargetPosition() const
+{
+    return mRestrictedTargetPosition;
+}
+const osg::Vec3f& Data::OsgNode::restrictedTargetPositionConstRef() const
+{
+    return mRestrictedTargetPosition;
+}
+void Data::OsgNode::setRestrictedTargetPosition( const osg::Vec3f& position )
+{
+    mRestrictedTargetPosition = position;
 }
